dm-io: reserve just one bioset structure
dm-io: reserve just one bioset structure

The number 16 was "magic" and meaningless. This patch changes it to reserve just
one "struct bio", which is the minimum required.

If there is no memory pressure, the code can allocate arbitrary number of bios.

If there is memory pressure, the code can allocate just one bio. The code in
"do_region" contains a loop, it allocates a bio, fills it and submits it. If we
have just one bio available, the code wouldn't deadlock, it would just wait
with the allocation of next bio until the previous bio finished.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>