[block-manager] Some reworking of recycle_block.
recycle_block now expects to be called with the spinlock held.  It
does drop it when waiting for io, but has retained it before
returning.  This prevents a race condition when using the returned
block.

The bulk of the patch is just some code movement, the validator
checking now occurs after the lock is aquired.