[block-manager, thin-metadata] Make the bm kmemcache name unique to the pool.
ATM the kmem_cache identifier just includes the major/minor of the
device that the bm is looking at.  This means you can't create 2 bms
that point to the same device without generating a nasty stack trace.

This patch makes the name unique to the pool.  It's a quick hack to
allow me to keep on testing.  Could someone who cares about these
kmemcaches decide on a proper solution please.