Buffered I/O interface.
This interface allows you to do I/O on device and act as a cache,
holding recently read blocks in memory and performing delayed writes.

We don't use buffer cache or page cache already present in the kernel, because:
* both buffer cache and page cache can't handle block size larger than one page
* both buffer cache and page cache allocate memory to perform reads, this could
  lead to deadlocks

Currently, we use a fraction of available memory for cache, the settings can
be changed in /sys/module/dm_store_mikulas/parameters/dm_bufio_cache_size file.

If mm people provide some callbacks, we could install a callback to free the
cache in case of memory pressure.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>