dm_btree: fix rebalancing of 3 nodes after remove
When we remove an entry from a node we sometimes rebalance with it's
neighbours.  This wasn't being done correctly.  This code is barely
used currently; only when you delete a thin device, and then only if
you have hundreds of them in the same pool.  Once we have discard
support, which removes mappings this will be used much more heavily.

Signed-off-by: Joe Thornber <ejt@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>