dm-thin: don't use the bi_next field for the holder of a cell.
When a cell is created the bio that triggered creation (the holder)
was added to the same bio list as subsequent bios.  In some cases we
let the holder continue to lower devices.  If those lower devices use
the bi_next field there will be trouble ...

This change keeps the holder in a new, separate field of the cell.

This also has the effect of simplifying some code that had to work out
which bio was the holder.

Signed-off-by: Joe Thornber <ejt@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>