dm thin: fix memory leak of singleton bio-prison cell
Add missing mempool_free() to __cell_release_singleton().

This is a pretty significant leak that will accumulate to over 2GB of
leaked memory just from running the full thinp-test-suite.  Leak was
known to exist but the kmalloc_pool which was used for prison cell
allocation caused all leaked memory to be attributed to the "size-128"
cache.  kmemleak proved useful in identifying the source of this
illusive leak.

Signed-off-by: Mike Snitzer <snitzer@redhat.com>