dm-cache: Check for IO pattern being random before caching
Somewhere we lost the actual check for randomness of IO stream and started
caching everything. That made this cache-noseq branch same as thin-dev
branch in performance.

This time make the check even more stringent and disable caching even if
it is cheap copy. Reason being that there is a cost to evict the cache
entries when actual random IO happens and we want to cache that.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>