dm thin: allow discards even if block size is not a power of 2
The block layer assumes discard_granularity is a power of 2, so changes
are needed to allow discards when a pool is using a block size that is
not a power of 2.  This patch depends on commit c6e666345e1b79c62b
("block: split discard into aligned requests").

If discard passdown is enabled: use the data device's
discard_granularity but make sure block size is a multiple of it.  Also
make sure the data device's discard_max_sectors is not smaller than the
pool's block size.

If discard passdown is disabled: use a discard_granularity of
DATA_DEV_BLOCK_SIZE_MIN_SECTORS if block size isn't a power of 2.  The
pool's block size is a multiple of DATA_DEV_BLOCK_SIZE_MIN_SECTORS
(which is a power of 2).

Signed-off-by: Mike Snitzer <snitzer@redhat.com>