dm thin: use proper pool_features structure when establishing discard limits
pool_io_hints() is executed _before_ the pool has been resumed; so pt->pf
must be used instead of pool->pf (long standing bug).  BUT when
disabling discard_passdown we cannot use pt->pf to do so because that
would result in a thin-pool table line conflict (which lvm2 would balk
at); so introduce a 'discard_passdown_override' member in the
pool_features structure (use it to reestablish the
pool->pf.discard_passdown after pt->pf is handed to pool->pf).

Not seeing how we avoid this nuanced complexity...