dm cache: rename cache_c structure to cache
Also rename 'c' variables to 'cache'.

Paves way for 2 classes of cache objects should we want to go that route
in the future (cache and cache_c, ala pool and pool_c).  Even after such
a split the bulk of the code will be dealing with 'struct cache', so
that is the reason for the pre-emptive switch to 'struct cache' and
s/c/cache/.