[dm-cache] policy_lookup.
The get_num_duplicates target method was calling policy_map, which was
treating this as a hit (eg, promoting to the pre-cache etc).  The new
policy_lookup method has no side effects.

In addition we now assume it's a writethrough io, requiring 2
duplicates, if the lookup returns an error such as -EWOULDBLOCK.  I
think this may have been the source of the corruption that Mike was
seeing yesterday.