[dm-thin] Fix deadlock when io and discards are simultaneously sent to the same block.
The all_io_entry for a bio must be incremented whilst the bio_prison
cell is held for both the virtual and physical blocks.  The cell
locking wasn't occuring.

This patch fixes this.  Since the bio_prison_cells are only held
briefly we don't bother allocating them from the mempool, instead
using a couple in the struct thin_c protected by a spinlock (if they
were smaller we'd just put them on the stack).