dm: introduce per-bio data
Introduce a field per_bio_data_size in struct dm_target.

Targets can set this field in the constructor. If a target sets this
field to a non-zero value, "per_bio_data_size" bytes of auxiliary data are
allocated for each bio submitted to the target. These data can be used
for any purpose by the target.

Per-bio data is accessed with dm_per_bio_data. The
argument data_size must be the same as the value per_bio_data_size in
dm_target.

If the target has a pointer to per_bio_data, it can get a pointer to
the bio with dm_bio_from_per_bio_data() function (data_size must be the
same as the value passed to dm_per_bio_data).

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>

FIXME Add __ to dm_target_io because it's exposed for optimisation and shouldn't be really?
FIXME   Or just accept that it's now exposed and can we use it in any useful way?
FIXME Overlap with 'info': can that field be removed?  (Yes, now in later patches)
FIXME Can we avoid having to re-specify per-bio data size when accessing it?
FIXME Alignment of data