dm cache metadata: add policy hint_size awareness
If policy's hint_size is 0 the the hint array isn't initialized/used.

policy_hint_size is added to the superblock but it doesn't _need_ to be
(I added it for symmetry with other variable passing between methods).

Hopefully this work gets us a bit closer to supporting variable hint_size.