dm cache metadata: improve policy checking
Introduce policy_unchanged() and have it check that the policy name,
major version and hint size are all unchanged.

If the policy's major version has changed then it is assumed that the
hints are no longer compatible with previous versions of the same
policy.

Pass policy structure around rather than just policy_name.

Signed-off-by: Mike Snitzer <snitzer@redhat.com>