[dm-thin] Be a bit cleverer about calculating the metadata threshold.
When we get low on free metadata blocks we trigger a dm event,
allowing userland to resize the metadata device.

Previously this threshold was a percentage of the metadata device.
Now it's the minimum of 4M or a percentage.  Should stop us triggering
too early for very large metadata device.