[dm-cache] Performance improvement in mq policy: do not update the stats in map if you're going to return -EWOULDBLOCK
The io thread calls the policy's map method.  If it would like to
migrate it returns -EWOULDBLOCK, the io gets handed over to a worker
thread which handles the migration.  The worker thread calls map
again, to check the migration is still needed (avoids a race among
other things).  *BUT* we're probably still in the same tick period, so
if the stats have been updated by the previous map call, the migration
will no longer be requested.

This was why micro benchmarks that hammered IO to a single block were
failing to cause a promotion.