[dm-cache] Introduce three new tunables to the mq policy
Internally the mq policy maintains a promotion threshold variable.  If
the count of a block not in the cache goes above this threshold it
gets promoted to the cache.  These tunables allow you to tweak the
promotion threshold by adding a small value.  These adjustments depend
on the io type:

   read_promote_adjustment:    READ io, default 4
   write_promote_adjustment:   WRITE io, default 8
   discard_promote_adjustment: READ/WRITE io to a discarded block, default 1

If you're trying to quickly warm a new cache device you may wish to
reduce these to encourage promotion.  Remember to switch them back to
their defaults after the cache fills though.