dm: change sector_t len to unsigned
It is impossible to create bios with 2^23 or more sectors (the size is
stored as a 32-bit byte count in the bio). So we convert some sector_t
values to unsigned integers.

It is needed for the next patch that replaces the values with pointers, so
the size of the integer must match.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>