dm: remove num_write_bios
A DM target can set the function num_write_bios - DM core will issue
this callback to ask the target how many bios it wants to receive.  But
num_write_bios is unused, so remove it.

It was intended for the dm-cache target, but it is not usable due to a
race condition (see the description of commit e2e74d617, "dm cache: fix
race in writethrough implementation").

Note that the for loop in __clone_and_map_data_bio() was deliberately
left because it will be used by the next commit ("dm: introduce
dm_ask_for_duplicate_bios").

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>