dm: introduce dm_ask_for_duplicate_bios
This function can be used if the target needs to receive a duplicate of
the current bio.

In dm_ask_for_duplicate_bios, the value of the dm_target_io's 'num_bios'
pointer is incremented to add @num_duplicates to the outer control loops
in the bio mapping callers (e.g. __send_duplicate_bios and
__clone_and_map_data_bio).

dm_ask_for_duplicate_bios() will first be used by the DM cache target to
improve write-through performance.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>