[dm-thin] transition from out-of-data-space mode if blocks are released
Discard bios, and thin deletiion have the potential to release data
blocks.  If we are in out-of-data-space mode, then we can transition
back to full write mode.

The correct time to do this is just after the commit.  Before the
commit the space maps will not let us reuse the data blocks in case
there's a rollback following power failure.