dm-cache: Allow sharing of cache-metadata objects.
If a dm table is reloaded when the device is not suspended (normal
procedure for LVM2), then there will be two dm-bufio objects that can
diverge.

This patch uses reference counting and a global list of cache metadata
objects to ensure there is only one metadata object per metadata
device.