dm-cache: Fix race when issuing a POLICY_REPLACE operation.
There is a race between a policy deciding to replace a cache entry,
the core target writing back any dirty data from this block, and other
IO threads doing IO to the same block.

This sort of problem is avoided most of the time by the core target
grabbing a bio prison cell before making the request to the policy.
But for a demotion the core target doesn't know which block will be
demoted, so can't do this in advance.

This patch introduces a callback to the policy interface that allows
the policy to grab the cell for the core target.

[Mike move to start of set please]