dm.c: Fix coercion bug in dm_merge_bvec()
dm_merge_bvec was originally added in f6fccb.  In this patch a value
in sectors is converted to bytes via a << 9, and then assigned to an
int.  This code made assumptions about the value of BIO_MAX_SECTORS.

A later patch, 148e51, removed the use of BIO_MAX_SECTORS.  At this
point the coercion resulted in a zero value.  The upshot being
dm_merge_bvec would only allow a single page to be added to a bio.

This patch is an interim patch.  The full solution needs to change
the max_size parameter in the dm_target merge methods to take a
sector_t (containing sectors, not bytes).