dm-thin: metadata space map sometimes leaked metadata blocks on resize
The metadata space map has a simplified 'bootstrap' mode that is
operational when extending the space maps.  Whilst in this mode it's
possible for some dec operations to become queued (eg, as a result of
shadowing one of the bitmap indexes).  These decrements were not being
applied when switching out of bootstrap mode.

The effect of this bug was the leaking of a 4k metadata block.  This
is detected by the latest version of thin_check as a non fatal error.