[dm-btree, dm-thin] Add ref counting ops for the leaves of top level btrees.
When using nested btrees, the top leaves of the top levels contain
block addresses for the root of the next tree down.  If we shadow a
shared one of these leaf nodes the leaf values (sub tree roots) should
be incremented accordingly.

This is only an issue if there is metadata sharing in the top levels.
Which only occurs if metadata snapshots are being used.  And could
result in a block from the metadata snap being reused early, thus
corrupting the metadata snap.