[dm-btree] Fix bug when rebalancing nodes after removal
The redistribute3 function takes 3 btree nodes and shares out the entries
evenly between them.  If the three nodes in total contained
(MAX_ENTRIES * 3) - 1 entries between them then this was erroneously getting
rebalanced as (MAX_ENTRIES - 1) on the left and right, and (MAX_ENTRIES + 1) in
the center.

This patch is more careful about calculating the target nr entries for the left
and right nodes.

Unit tested in userspace using this program:

https://github.com/jthornber/redistribute3-test/blob/master/redistribute3_t.c