[dm-thin] Delete thin devices in the background.
dm_btree_del_background() does most of it's work with a read_lock
held, and only takes the write lock every now and then to do a batch
of ref count decrements for freed blocks.  This should allow other IO
to the pool to proceed in parallel with the removal (which can take
some time for v. large thin devices).  If there's a crash during
deletion of a thin device then unfreed blocks will be leaked, an issue
that can be identified by thin_check, and fixed with thin_repair.