[dm-cache] Make the mq policy an alias for smq
smq seems to be performing better than the old mq policy in all
situations, as well as using a quarter of the memory.

This patch makes 'mq' and alias for 'smq' when choosing a cache
policy.  The tunables that were present for the old mq are faked, and
have no effect.  mq should be considered deprecated now.