[dm-cache] Use dm_array_mutate() to speed up dirty bitset writing.
eg, Brings writing for a v. large cache down from ~160 seconds to ~40
seconds.  Still too slow because we're mutating the whole mapping
array.  This also requires a lot of spare metadata space to hold this
huge transaction.  I think we're going to have to split the dirty
bitset out (like the discard bitset).