[dm-thin] Fix race when discarding regions
A race condition exists between one thin device discarding a block and
another provisioning it.

The discard passdown was being issued after the unmapping, which meant
the block could be reprovisioned whilst the passdown discards are
still in flight.

We can only identify unshared blocks (safe to do a passdown to) once
they're unmapped and their ref count hits zero.  So now we unmap,
issue passdowns, and immediately increment regions that have been
passed down (this is safe because allocation occurs on the same
thread).  We then decrement ref counts once the passdown IO is
complete.