[dm-cache-policy-smq] Distribute entries to random levels when switching to smq.
For smq the 32 bit 'hint' stores the multiqueue level that the entry
should be stored in.  If a different policy has been used previously,
and then switched to smq the hints will be invalid.  In which case we
used to put all entries in the bottom level of the multiqueue, and
then redistribute.  Redistribution is faster if we put entries with
invalid hints in random levels initially.