dm cache: remove "invalidate_cblocks" message support
The "invalidate_cblocks" message was previously provided for a very
specialized use-case that required invalidating portions pf the cache
due to rolling back a hardware vendor's snapshot.  The new dm-cache code
doesn't lend itself to easily supporting "invalidate_cblocks" but given
that it is very likely unused anyway just remove all the code and error
out with -EOPNOTSUPP for the foreseeable future.

Signed-off-by: Mike Snitzer <snitzer@redhat.com>